changed the file structure
